pyproject.toml
[project]
name = "agent-inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.19.0",
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
